requirements.txt
python-telegram-bot~=21.10
youtube_transcript_api~=1.0
